pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "linkora"
version = "0.1.1"
description = "Local Knowledge Network — AI-native research cli powered by local-first architecture"

requires-python = ">=3.12"
license = "MIT"
authors = [
    # The original project creator
    { name = "Zimo Liao", email = "zimoliao@mail.ustc.edu.cn" },
    { name = "nostalgia", email = "lv15623448685@gmail.com" },
]
keywords = [
    "academic",
    "literature",
    "research",
    "knowledge-base",
    "semantic-search",
    "mcp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Indexing",
]
dependencies = ["requests>=2.28", "pyyaml>=6.0", "tenacity>=9.1.4"]

[project.urls]
Homepage = "https://github.com/lvyuemeng/linkora"
Repository = "https://github.com/lvyuemeng/linkora"

[project.optional-dependencies]
embed = [
    "sentence-transformers>=3.0",
    "numpy>=1.24",
    "faiss-cpu>=1.7",
    "modelscope>=1.10",
]
topics = ["linkora[embed]", "bertopic>=0.16", "pandas>=2.0"]
import = [
    "endnote-utils>=1.0",
    "pyzotero>=1.5",
    "rapidfuzz>=3.14.3",
    "tqdm>=4.67.3",
]
full = [
    "linkora[embed]",
    "linkora[topics]",
    "linkora[import]",
    "curl-cffi>=0.5",
]

[project.scripts]
linkora = "linkora.cli:main"

[dependency-groups]
dev = ["pytest>=7.0", "ruff>=0.15.5", "ty>=0.0.23"]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = ["-v", "--tb=short"]
markers = [
    "unit: Unit tests (pure functions)",
    "integration: Integration tests (multiple modules)",
]
